+++ pkg/storegateway/threadpool/threadpool.go
@@ -3,12 +3,14 @@
 // Provenance-includes-license: MIT
 // Provenance-includes-copyright: Christian Grabowski
 
-package indexheader
+package threadpool
 
 import (
+	"context"
 	"errors"
 	"time"
 
+	"github.com/grafana/dskit/services"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 )
@@ -21,14 +23,14 @@ const (
 var ErrPoolStopped = errors.New("thread pool has been stopped")
 
 type Threadpool struct {
+	services.Service
+
 	// pool is used for callers to acquire and return threads, blocking when they are all in use.
 	pool chan *osThread
 	// threads is used to perform operations on all threads at once (such as stopping and shutting down).
 	threads []*osThread
 	// stopping is closed when calling code wants the threadpool to shut down.
 	stopping chan struct{}
-	// stopped is closed once all threads have stopped running.
-	stopped chan struct{}
 
 	timing *prometheus.HistogramVec
 	tasks  prometheus.Gauge
@@ -36,30 +38,26 @@ type Threadpool struct {
 
 // NewThreadPool creates a new instance of ThreadPool that runs tasks in a pool
 // of dedicated OS threads, meaning no other goroutines are scheduled on the threads.
-// If num is less than or equal to zero, nil is returned.
-func NewThreadPool(num int, reg prometheus.Registerer) *Threadpool {
-	if num <= 0 {
-		return nil
-	}
-
+// If num is zero, no threads will be started and all calls to Execute will run in
+// the goroutine of the caller.
+func NewThreadPool(num uint, reg prometheus.Registerer) *Threadpool {
 	tp := &Threadpool{
 		pool:     make(chan *osThread, num),
 		threads:  make([]*osThread, num),
 		stopping: make(chan struct{}),
-		stopped:  make(chan struct{}),
 
 		timing: promauto.With(reg).NewHistogramVec(prometheus.HistogramOpts{
-			Name:    "cortex_bucket_store_indexheader_thread_pool_seconds",
-			Help:    "Amount of time spent performing index header operations on a dedicated thread",
+			Name:    "cortex_storegateway_thread_pool_seconds",
+			Help:    "Amount of time spent performing operations on a dedicated thread",
 			Buckets: []float64{0.001, 0.01, 0.1, 1, 10},
 		}, []string{"stage"}),
 		tasks: promauto.With(reg).NewGauge(prometheus.GaugeOpts{
-			Name: "cortex_bucket_store_indexheader_thread_pool_tasks",
-			Help: "Number of index header operations currently executing",
+			Name: "cortex_storegateway_thread_pool_tasks",
+			Help: "Number of thread pool operations currently executing",
 		}),
 	}
 
-	for i := 0; i < num; i++ {
+	for i := uint(0); i < num; i++ {
 		t := newOSThread(tp.stopping)
 		t.start()
 
@@ -71,38 +69,33 @@ func NewThreadPool(num int, reg prometheus.Registerer) *Threadpool {
 		tp.pool <- t
 	}
 
+	tp.Service = services.NewBasicService(nil, tp.run, nil)
 	return tp
 }
 
-func (t *Threadpool) run() {
-	defer close(t.stopped)
+func (t *Threadpool) run(ctx context.Context) error {
+	<-ctx.Done()
 
-	// The t.stopping channel is never written so this blocks until the channel is
-	// closed at which point the threadpool is shutting down, so we want to stop
-	// each of the expected threads in it.
-	<-t.stopping
-
+	// Close that channel that we've given to our threads to indicate that they
+	// should stop accepting new work and end.
+	close(t.stopping)
 	// Wait for all threads, regardless if they are "in" the pool or being used to
 	// run caller code. The avoids race conditions where threads are removed and
 	// added back to the pool while we are trying to stop all of them.
 	for _, thread := range t.threads {
 		thread.join()
 	}
-}
 
-func (t *Threadpool) Start() {
-	go t.run()
-}
-
-func (t *Threadpool) StopAndWait() {
-	// Indicate to all threads that they should stop, then wait for them to do so
-	// by trying to read from a channel that will be closed when all threads have
-	// finally stopped.
-	close(t.stopping)
-	<-t.stopped
+	return nil
 }
 
 func (t *Threadpool) Execute(fn func() (interface{}, error)) (interface{}, error) {
+	// No threads to run the closure in. This is how operators disable use of the threadpool
+	// so just run the closure in the caller's goroutine in that case.
+	if len(t.threads) == 0 {
+		return fn()
+	}
+
 	start := time.Now()
 
 	select {